Add files via upload
Two main components are: the primary and secondary systems, in the primary system which has the uttermost priority, is the data acquisition by sEMG. Once the EMG signal is acquired from the kit (which already rectifies and smooths it as the internal function of the kit), it is fed into the microcontroller. The microcontroller has programmed filters that are required for the acquisition of useful data and the removal of noise and motion artifacts from the signal.

Secondary input comprises state information from the two limit switches mounted on the motor pulley. At any instant, either one of them gets HIGH. The two limit switches allow us two know the limited two joint positions: fully flexed or extended and at any moment, either of them is achieved.

Now we have two sets of data, the joint position from the limit switches and the required EMG signal after filtering, but how do we know what the muscle intends to do, how much force is it exerting, and what position it wants the joint to move? this is done by feeding the EMG signal into the microcontroller which calculates the intended position by the Threshold Difference method. After the desired output, the microcontroller commands the motor unit to move the elbow joint. Feedback from the limit switches is again compared and this process keeps on iterating.